Change type of IUIAutomationStructureChangedEventHandler.rf to uintptr
errro after this commit:
```
C:\Users\hnakamur\w32uiautomation\examples\runcalc>go run main.go
Exception 0xc0000005 0x8 0x1 0x1
PC=0x1

runtime.cgocall(0x460440, 0x599800)
        c:/go/src/runtime/cgocall.go:157 +0x4a fp=0xc00011bd18 sp=0xc00011bce0 pc=0x403aca
syscall.SyscallN(0x599540?, {0xc00011bdb0?, 0x1?, 0xc00011bdd8?})
        c:/go/src/runtime/syscall_windows.go:538 +0x109 fp=0xc00011bd90 sp=0xc00011bd18 pc=0x45b929
syscall.Syscall6(0xc0000125c0?, 0x1?, 0x1?, 0xc00011beb0?, 0xc000088040?, 0x20714bd37d0?, 0x20714c132a0?, 0xc000088060?)
        c:/go/src/runtime/syscall_windows.go:482 +0x50 fp=0xc00011bdf0 sp=0xc00011bd90 pc=0x45b5b0
github.com/hnakamur/w32uiautomation.findFirst(0x0?, 0x0?, 0x0?)
        C:/Users/hnakamur/w32uiautomation/iuiautomationelement.go:164 +0x5b fp=0xc00011be68 sp=0xc00011bdf0 pc=0x4ba49b
github.com/hnakamur/w32uiautomation.(*IUIAutomationElement).FindFirst(...)
        C:/Users/hnakamur/w32uiautomation/iuiautomationelement.go:118
github.com/hnakamur/w32uiautomation.WaitFindFirst(0x8?, 0x20714c15438?, 0x0?, 0x0?)
        C:/Users/hnakamur/w32uiautomation/waitfind.go:12 +0x57 fp=0xc00011beb0 sp=0xc00011be68 pc=0x4ba9f7
main.runCalc()
        C:/Users/hnakamur/w32uiautomation/examples/runcalc/main.go:42 +0x145 fp=0xc00011bf50 sp=0xc00011beb0 pc=0x4bada5
main.main()
        C:/Users/hnakamur/w32uiautomation/examples/runcalc/main.go:100 +0x45 fp=0xc00011bf80 sp=0xc00011bf50 pc=0x4bb185
runtime.main()
        c:/go/src/runtime/proc.go:250 +0x1fe fp=0xc00011bfe0 sp=0xc00011bf80 pc=0x43641e
runtime.goexit()
        c:/go/src/runtime/asm_amd64.s:1571 +0x1 fp=0xc00011bfe8 sp=0xc00011bfe0 pc=0x45ebc1
rax     0x1
rbx     0x20714c673d0
rcx     0xc00011be48
rdi     0x7ffe905eb110
rsi     0xbc717fc100
rbp     0xbc717fc130
rsp     0xbc717fc068
r8      0x0
r9      0x0
r10     0x0
r11     0x0
r12     0x1
r13     0x20714c6cca8
r14     0x1
r15     0xdeaddeaddeaddead
rip     0x1
rflags  0x10206
cs      0x33
fs      0x53
gs      0x2b
exit status 2
```

+++ iuiautomationstructurechangedeventhandler.go
@@ -40,7 +40,7 @@ func (t StructureChangeType) ToString() string {
 
 type IUIAutomationStructureChangedEventHandler struct {
 	ole.IUnknown
-	ref int32
+	ref uintptr
 }
 
 type IUIAutomationStructureChangedEventHandlerVtbl struct {
@@ -70,13 +70,13 @@ func structureChangedEventHandler_queryInterface(this *ole.IUnknown, iid *ole.GU
 	return ole.E_NOINTERFACE
 }
 
-func structureChangedEventHandler_addRef(this *ole.IUnknown) int32 {
+func structureChangedEventHandler_addRef(this *ole.IUnknown) uintptr {
 	pthis := (*IUIAutomationStructureChangedEventHandler)(unsafe.Pointer(this))
 	pthis.ref++
 	return pthis.ref
 }
 
-func structureChangedEventHandler_release(this *ole.IUnknown) int32 {
+func structureChangedEventHandler_release(this *ole.IUnknown) uintptr {
 	pthis := (*IUIAutomationStructureChangedEventHandler)(unsafe.Pointer(this))
 	pthis.ref--
 	return pthis.ref
